### 2017-03-07 #200 @whedon
[PRE REVIEW]: Virtual Bumblebees
**Submitting author:** @howardjp (<a href="http://orcid.org/0000-0003-4530-1547">James P. Howard, II</a>)
**Repository:** <a href="https://github.com/howardjp/bumblebees" target ="_blank">https://github.com/howardjp/bumblebees</a>
**Version:** 1.1

Thanks for submitting your paper to JOSS howardjp. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
