### 2017-03-12 #200 @arfon
Hi @howardjp. It's not entirely clear what this submission is? Could you tell me a little more about the research application of this work?
