### 2017-03-13 #200 @howardjp
That's a good question, @arfon.  The environment is an implementation of a specific cellular automata ruleset.  This is [artifical life](https://en.wikipedia.org/wiki/Artificial_life).  In particular, it is a variant of the Langton Ant (see, e.g., [Gajardo, Moreira, and Goles, 2002](http://www.sciencedirect.com/science/article/pii/S0166218X00003346)).  In addition, one could implement the Langton Ant within Virtual Bumblebees through careful construction of a starting grid.

Here's a copy of the [presentation I gave at MathFest 2015](https://1drv.ms/p/s!AoC3AjtvhtrEhi94FPzEv-ZTXBZ7).  Be warned, it is 160M due to embedded videos.
