### 2017-04-24 #200 @arfon
> @arfon I wanted to jump in and see if this had completely stalled or if I could help rekickstart it somehow?

Apologies for dropping the ball @howardjp. I'm struggling to figure out who might be a suitable editor for this submission.

@openjournals/joss-editors - anyone willing to volunteer to edit this submission?
