### 2017-04-24 #200 @howardjp
It's all good and we've all been there.  Just wanted to see if I could give it a nudge.  Totally unrelated, @arfon, but I spoke with one of your colleagues at NerdNite Baltimore last week.  It's a small world.
