### 2017-04-25 #200 @pjotrp
It is a nifty visualisation. I.e., pretty cool. I do, however, think that this submission is going to stretch our scientific application rule. @howardjp unlike the papers you refer to you don't give an explanation of what you actually have implemented and how it differs from the ant? I think that would be incredibly helpful. Your code is short and concise, but not exactly explanatory.

Also can you run through the list of checkboxes of another submission, such as https://github.com/openjournals/joss-reviews/issues/248 and see if we can actually tick the boxes.

Finally, do consider that your your scientific credibility is attached to your submission and may be visible for eternity ;)
