### 2017-04-25 #200 @howardjp
@pjotrp I think a lot of the checkboxes can be ticked with better documentation, so I will work on that and ping you again in a few days.

As for scientific credibility, in a field that only progresses when someone has made a mistake, it is important not to take oneself too seriously ;)
