### 2017-04-27 #200 @howardjp
@pjotrp Okay, I added a [README.md file to the develop branch](https://github.com/howardjp/bumblebees/blob/develop/README.md).  This adds a lot of context and explanation for

1. Why I created this,
2. How to use it, and
3. How it works.

Of questions on this checklist, this will resolve almost all of them.  Some really don't apply (API documentation, though I do describe the internals, which can be helpful to someone wanting to make changes).  So with this in mind, please take another look and I can take another stab at getting it up to snuff.

Thanks! James
