### 2017-05-05 #200 @howardjp
Hi, sorry for my own delay.  The final week of class is dense, to say the least.  As it is, not really.  My education is in public policy.  I literally wrote Bumblebees as a one-off while sitting at my son's chess tournament, as I described in a [blog post](https://jameshoward.us/2015/03/14/introducing-my-bumblebees/), based on something I was messing with 20-odd years ago.  I just wanted someone else to find it and do something interesting with it. :)
