### 2017-05-08 #256 @whedon
[REVIEW]: Virtual Bumblebees
**Submitting author:** @howardjp (<a href="http://orcid.org/0000-0003-4530-1547">James P. Howard, II</a>)
**Repository:** <a href="https://github.com/howardjp/bumblebees" target ="_blank">https://github.com/howardjp/bumblebees</a>
**Version:** 1.1

- [x] **Authorship:** Has the submitting author (howardjp) made major contributions to the software?
