### 2017-05-09 #256 @pjotrp
@howardjp: I started review. Thank you for your contribution to JOSS. Two minor comments:

1. Please merge your development branch with master
2. Check github recognizes the LICENSE - in the repo header

When that is done I'll tick the final boxes.
