### 2017-05-09 #256 @howardjp
1. Check, now version 1.2.1
2. Wow, I just want to say, that was an entirely nontrivial operation
