### 2017-05-10 #256 @pjotrp
Your software is compliant ;). One last niggle, can you make a release of the software under version 1.2.1? Once that is done we can in principle assign a DOI.
