### 2017-05-11 #256 @pjotrp
Thank you @howardjp. The release should match the paper version.

@arfon I think we can go ahead and accept this submission.
