### 2017-05-11 #256 @arfon
Many thanks for your review @pjotrp ✨

@howardjp - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00256 ⚡️ 🚀 💥
